+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Breaking Changes
 * `NavigationMapViewDelegate` and `RouteMapViewControllerDelegate`: `navigationMapView(_:didTap:)` is now `navigationMapView(_:didSelect:)` [#1063](https://github.com/mapbox/mapbox-navigation-ios/pull/1063)
 
+#### User Interface
+* `StepsViewController` 's convienence initalizer (`StepsViewController.init(routeProgress:)`) is now public.
+
 ## v0.14.0 (February 22, 2018)
 
 #### Breaking Changes